api/requirements.txt
transformers>=4.40.0
librosa>=0.10.0
opencv-python-headless>=4.9.0
httpx>=0.27.0
anthropic>=0.28.0
google-generativeai>=0.7.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
